app/devsu-demo-devops-python/requirements.txt
django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
whitenoise==6.5.0
psycopg2-binary==2.9.9